fix : 1) store 정리(게시물, 댓글) 및 Redux-persist 설정, 2) 대댓글 작성 시 자동 리스트 추가, 3) 알람 시간 표시 수정, 4) 투명 배경 나오는 페이지 수정
