+++ package.json
@@ -8,7 +8,6 @@
     "@testing-library/jest-dom": "^5.14.1",
     "@testing-library/react": "^12.0.0",
     "@testing-library/user-event": "^13.2.1",
-    "@trendyol-js/react-carousel": "^2.0.1",
     "aws-sdk": "^2.1092.0",
     "firebase": "^9.6.10",
     "history": "^5.3.0",
@@ -26,6 +25,8 @@
     "react-router-dom": "6.2.1",
     "react-scripts": "5.0.0",
     "react-select": "^5.2.2",
+    "redux-persist": "^6.0.0",
+    "redux-persist-expo-securestore": "^2.0.0",
     "simple-peer": "^9.11.1",
     "sockjs-client": "^1.5.2",
     "stompjs": "^2.3.3",
