+++ src/modules/comments.js
@@ -1,16 +1,7 @@
-import {
-  createAsyncThunk,
-  createSlice,
-  createSelector,
-} from '@reduxjs/toolkit';
+import { createAsyncThunk, createSlice } from '@reduxjs/toolkit';
 import apis from '../apis/apis';
 
 const commentListInitialState = {
-  data: [],
-  status: 'idle',
-};
-
-const replyCommentListInitialState = {
   data: [],
   status: 'idle',
 };
@@ -27,18 +18,6 @@ export const getCommentListAsync = createAsyncThunk(
   },
 );
 
-export const getReplyCommentListAsync = createAsyncThunk(
-  'comments/getReplyCommentList',
-  async (commentId, thunkAPI) => {
-    try {
-      const response = await apis.getReplyCommentListByComment(commentId);
-      return { commentId: commentId, list: response.data };
-    } catch (e) {
-      return thunkAPI.rejectWithValue(await e.response.data);
-    }
-  },
-);
-
 export const createCommentAsync = createAsyncThunk(
   'comments/createComment',
   async (commentInfo, thunkAPI) => {
@@ -49,21 +28,6 @@ export const createCommentAsync = createAsyncThunk(
       .catch((error) => {
         return thunkAPI.rejectWithValue();
       });
-  },
-);
-
-export const createReplyCommentAsync = createAsyncThunk(
-  'comments/createComment',
-  async (replyCommentInfo, thunkAPI) => {
-    const { commentId, replyContent } = replyCommentInfo;
-    console.log(commentId);
-    await apis
-      .createReplyComment(commentId, replyContent)
-      .then((response) =>
-        thunkAPI.dispatch(
-          addReplyComment({ commentId: commentId, list: response.data }),
-        ),
-      );
   },
 );
 
@@ -97,31 +61,6 @@ export const recommendCommentAsync = createAsyncThunk(
             thunkAPI.dispatch(increaseRecommendCount(commentId));
           } else {
             thunkAPI.dispatch(decreaseRecommendCount(commentId));
-          }
-        })
-        .catch((error) => {
-          if (error) {
-            window.alert('잘못된 추천 요청입니다.');
-          }
-          return thunkAPI.rejectWithValue();
-        });
-    }
-  },
-);
-
-export const recommendReplyCommentAsync = createAsyncThunk(
-  'comments/recommendReplyCommentCount',
-  async (ids, thunkAPI) => {
-    const { commentId, replyId } = ids;
-    if (commentId && replyId) {
-      await apis
-        .recommendReplyComment(commentId, replyId)
-        .then((response) => {
-          console.log(response.data);
-          if (response.data.clicked) {
-            thunkAPI.dispatch(increaseReplyRecommendCount(replyId));
-          } else {
-            thunkAPI.dispatch(decreaseReplyRecommendCount(replyId));
           }
         })
         .catch((error) => {
@@ -172,57 +111,13 @@ export const commentListSlice = createSlice({
   },
 });
 
-export const replyCommentListSlice = createSlice({
-  name: 'replyCommentList',
-  initialState: replyCommentListInitialState,
-  reducers: {
-    clearReplyCommentList: (state) => {
-      state.status = 'idle';
-      state.data = [];
-    },
-    addReplyComment: (state, action) => {
-      console.log(action.payload);
-      state.data.push(action.payload);
-    },
-    increaseReplyRecommendCount: (state, action) => {
-      const replyIndex = state.data.findIndex((d) => {
-        return d.replyId === action.payload;
-      });
-      state.data[replyIndex].recommendCount += 1;
-    },
-    decreaseReplyRecommendCount: (state, action) => {
-      const replyIndex = state.data.findIndex((d) => {
-        return d.replyId === action.payload;
-      });
-      state.data[replyIndex].recommendCount -= 1;
-    },
-  },
-  extraReducers: {
-    [getReplyCommentListAsync.fulfilled]: (state, action) => {
-      state.status = 'success';
-      state.data.push(action.payload);
-    },
-    [getReplyCommentListAsync.pending]: (state) => {
-      state.status = 'loading';
-    },
-    [getReplyCommentListAsync.rejected]: (state) => {
-      state.status = 'failed';
-    },
-  },
-});
-
 export const {
   clearCommentList,
   addComment,
   increaseRecommendCount,
   decreaseRecommendCount,
 } = commentListSlice.actions;
-
-export const {
-  clearReplyCommentList,
-  addReplyComment,
-  increaseReplyRecommendCount,
-  decreaseReplyRecommendCount,
-} = replyCommentListSlice.actions;
 export const selectedCommentList = (state) => state.comments.data;
-export const selectedReplyCommentList = (state) => state.replyComments;
+export const selectedCommentCount = (state) =>
+  state.comments.data ? state.comments.data.length : 0;
+export default commentListSlice.reducer;
