+++ src/pages/boards/BoardList.jsx
@@ -5,11 +5,10 @@ import {
   clearBoardList,
   getBoardListAsync,
   getBoardListByCategoryAsync,
-  getMyBoardList,
   getMyBoardListAsync,
   getMyCommentListAsync,
   selectedBoardList,
-} from '../../modules/boards';
+} from 'modules/boards';
 
 import { Wrapper, Grid, Tile, Text } from 'components';
 import { useLocation, useParams } from 'react-router-dom';
